% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_functions.R
\name{addAnchorMatrix}
\alias{addAnchorMatrix}
\title{Add some prefixed sequence to existing energy matrix}
\usage{
addAnchorMatrix(energyMatrix, anchor, position = 1, height = 0.2)
}
\arguments{
\item{energyMatrix}{existing Energy Matrix}

\item{anchor}{The prefixed sequence}

\item{position}{The starting position of the prefixed sequence}

\item{height}{The height of prefixed sequence, which determines how it looks at energy logo.}
}
\value{
Energy matrix with some prefixed sequence at designated position
}
\description{
Add some prefixed sequence to existing energy matrix
}
\examples{
addAnchorMatrix(energyMatrix, anchor = "TAGC", position = 2, height = 0.25)
}
