% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{as.Coefficients}
\alias{as.Coefficients}
\title{Convert energy matrix to energy model coefficients
Encoding scheme: 3L+1}
\usage{
as.Coefficients(PEM)
}
\value{
3L+1 energy model coefficients
}
\description{
Convert energy matrix to energy model coefficients
Encoding scheme: 3L+1
}
\examples{
as.Coefficients(PEM)
}
