% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_functions.R
\name{as.PEM}
\alias{as.PEM}
\alias{as.PEM,matrix-method}
\alias{as.PEM,lm-method}
\title{Convert defined subject to position energy matrix}
\usage{
as.PEM(subject, ...)

\S4method{as.PEM}{matrix}(subject)

\S4method{as.PEM}{lm}(subject)
}
\description{
Convert defined subject to position energy matrix
}
\section{Methods (by class)}{
\itemize{
\item \code{matrix}: PFM

\item \code{lm}: energy model
}}

