% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{buildEnergyModel}
\alias{buildEnergyModel}
\title{Build energy model based on Sequence-Energy relationship}
\usage{
buildEnergyModel(data, encoding = "3L+1")
}
\arguments{
\item{data}{Data frame containing at least two variables -- Sequence and Energy}

\item{encoding}{The encoding scheme used to build model, including 3L+1, 4L+1, 5L+1.}
}
\value{
Linear model containing position-based energy coefficients, e.g., 1CA, 1CG, 1CT, 2CA, etc.
}
\description{
Build energy model based on Sequence-Energy relationship
}
\examples{
buildEnergyModel(data)
}
