% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_functions.R
\name{composeMotif}
\alias{composeMotif}
\title{Build composite motif based on two existing motifs with specified configurations}
\usage{
composeMotif(motifA, motifB, configuration = "F0F", type = "PEM")
}
\arguments{
\item{motifA}{The first motif (PEM by default)}

\item{motifB}{The second motif (PEM by default)}

\item{configuration}{"F0F" by default, i.e., Forward-0bp gap-Forward orientation}
}
\value{
compositeMotif constructed from motifA and motifB
}
\description{
Build composite motif based on two existing motifs with specified configurations
}
\examples{
buildCompositeMotif(SOX2.matrix, Oct4.matrix, configuration = "F0F")
}
