% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{countMismatch}
\alias{countMismatch}
\title{Count the number of mismatches of sequence(s) to some reference site}
\usage{
countMismatch(sequences, reference)
}
\arguments{
\item{sequences}{List of sequences.}

\item{reference}{Reference sequence.}
}
\value{
The number of mismatches between \code{sequences} and \code{reference}.
}
\description{
Count the number of mismatches of sequence(s) to some reference site
}
\examples{
countMismatch("GTAC", "GAAC")
countMismatch(c("GTAC", "GAAC"), "GAAC")
}
