% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_functions.R
\name{getEnergyMatrix}
\alias{getEnergyMatrix}
\alias{toPEM}
\title{Derive energy matrix from position-based energy model including parameters like 1CG, 1CT, 2CA, 3CG, etc}
\usage{
getEnergyMatrix(model)
}
\arguments{
\item{model}{position-based energy model, returned by buildEnergymodel function.}
}
\value{
Energy matrix
}
\description{
Derive energy matrix from position-based energy model including parameters like 1CG, 1CT, 2CA, 3CG, etc
}
\examples{
getEnergyMatrix(model)
}
