% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{highlightMismatch}
\alias{highlightMismatch}
\title{Highlight the mismatches of sequence(s) to some reference site with '-'}
\usage{
highlightMismatch(sequences, reference)
}
\arguments{
\item{sequences}{List of sequences.}

\item{reference}{Reference sequence.}
}
\value{
Sequences with those mismatched portions to reference changed to '-'
}
\description{
Highlight the mismatches of sequence(s) to some reference site with '-'
}
\examples{
highlightMismatch("GTAC", "GAAC")
highlightMismatch(c("GTAC", "GAAC"), "GAAC")
}
