% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plotEnergyLogo}
\alias{plotEnergyLogo}
\title{Visualize some energy matrix by plotting logo}
\usage{
plotEnergyLogo(PEM, col_scheme = c("default", "classic"))
}
\arguments{
\item{col_scheme}{The color scheme used to illustrate different nucleotides, which can be created by function ggseqlogo::make_col_scheme}
}
\value{
A ggplot object illustrating the energy logo.
}
\description{
Visualize some energy matrix by plotting logo
}
\examples{
plotEnergyLogo(PEM)
}
