% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{predictEnergy}
\alias{predictEnergy}
\title{Predict the binding energy of input sequences based on energy model}
\usage{
predictEnergy(sequences, model)
}
\arguments{
\item{sequences}{Input sequences}

\item{model}{Linear model containing position-based energy coefficients, e.g., 1CA, 1CG, 1CT, 2CA, etc.}
}
\value{
Predicted binding energy values of \code{sequences} based on \code{model}
}
\description{
Predict the binding energy of input sequences based on energy model
}
\examples{
predictEnergy(sequences, model)
}
