% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_functions.R
\name{reverseComplement}
\alias{reverseComplement}
\title{Reverse some existing energy matrix}
\usage{
reverseComplement(energyMatrix)
}
\value{
reversed energy matrix
}
\description{
Reverse some existing energy matrix
}
\examples{
reverseComplement(energyMatrix)
}
