% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{selectVariants}
\alias{selectVariants}
\title{Select sequence variants based on the max number of mismatches to some reference site}
\usage{
selectVariants(data, reference, maxMismatches = 1)
}
\arguments{
\item{data}{The data frame includes Sequence variable}

\item{reference}{Reference site}

\item{maxMismatches}{The max number of mismatches of each sequence to the reference site}
}
\value{
The data frame containing all sequences with no more than \code{maxMismatches} to the reference
}
\description{
Select sequence variants based on the max number of mismatches to some reference site
}
\examples{
selectVariants(data, "GTACGAC", maxMismatches=1)
}
