% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{SeqToByte}
\alias{SeqToByte}
\title{Convert DNA sequences to binary code based on various encoding schemes}
\usage{
SeqToByte(sequences, encoding = "3L+1")
}
\arguments{
\item{sequences}{List of sequences.}

\item{encoding}{The encoding scheme used to convert sequences to binary vector, including 3L+1, 4L+1, 5L+1. 3L+1 as default.}
}
\value{
Binary codes converted from sequences
}
\description{
Convert DNA sequences to binary code based on various encoding schemes
}
