% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{SeqToMethylByte}
\alias{SeqToMethylByte}
\title{Convert DNA sequences to binary code only including methylation-related bytes}
\usage{
SeqToMethylByte(sequences, encoding = "(3+2)L+1")
}
\arguments{
\item{sequences}{List of sequences.}

\item{encoding}{Either (3+1)L+1 or (3+1)L+1}
}
\value{
Binary codes converted from sequences, only M or W positions have non-zero bytes.
}
\description{
Convert DNA sequences to binary code only including methylation-related bytes
}
