% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{toCoeffs}
\alias{toCoeffs}
\title{Convert energy matrix to energy model coefficients
Encoding scheme: 3L+1}
\usage{
toCoeffs(energyMatrix)
}
\value{
energy model coefficients
}
\description{
Convert energy matrix to energy model coefficients
Encoding scheme: 3L+1
}
\examples{
toCoeffs("GTAC")
}
