% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{buildMethylationModel}
\alias{buildMethylationModel}
\title{Build energy model based on Sequence-Energy relationship}
\usage{
buildMethylationModel(data, encoding = "(3+2)L+1", withIntercept = FALSE)
}
\arguments{
\item{data}{Data frame containing at least two variables -- Sequence and Energy}
}
\value{
Linear model containing position-based energy coefficients, e.g., 1CA, 1CG, 1CT, 2CA, etc.
}
\description{
Build energy model based on Sequence-Energy relationship
}
\examples{
buildEnergyModel(data)
}
