% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_functions.R
\name{decomposeMotif}
\alias{decomposeMotif}
\title{Decompose some composite motif into two separate submotifs}
\usage{
decomposeMotif(compositeMotif, configuration = "F0F", type = "energy")
}
\arguments{
\item{configuration}{"F0F" by default, i.e., Forward-0bp gap-Forward orientation}

\item{motifA}{The first motif (energyMatrix by default)}

\item{motifB}{The second motif (energyMatrix by default)}
}
\value{
compositeMotif constructed from motifA and motifB
}
\description{
Decompose some composite motif into two separate submotifs
}
\examples{
buildCompositeMotif(SOX2.matrix, Oct4.matrix, configuration = "F0F")
}
