% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_functions.R
\name{reverseComplement}
\alias{reverseComplement}
\title{Reverse some existing position energy matrix}
\usage{
reverseComplement(PEM)
}
\arguments{
\item{matrix}{position energy matrix}
}
\value{
reversed energy matrix
}
\description{
Reverse some existing position energy matrix
}
\examples{
reverseComplement(PEM)
}
