% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{SeqToByte}
\alias{SeqToByte}
\title{Convert nucleotide sequence to binary code based on Bases->Codes correspondence}
\usage{
SeqToByte(sequences, encoding = "3L+1")
}
\arguments{
\item{sequences}{List of sequences.}

\item{Bases}{List of Bases to be mapped}

\item{Codes}{List of Binary codes corresponding to Bases}
}
\value{
Binary codes converted from sequences
}
\description{
Convert nucleotide sequence to binary code based on Bases->Codes correspondence
}
\examples{
SeqToByte("GTAC")
SeqToByte("GTAAN",
Bases=cbind('A', 'C', "G", "T", "N"),
Codes=cbind("0 1 0 ", "0 0 0 ", "1 0 0 ", "0 0 1 ", "0.25 0.25 0.25 "))
}
